NotionAPIのクラスをまとめた
