+++ Experience2Notion/Models/Notions/NotionPageBase.cs
@@ -1,0 +1,17 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions;
+public class NotionPageBase
+{
+    [JsonPropertyName("parent")]
+    public Parent Parent { get; set; } = new();
+
+    [JsonPropertyName("icon")]
+    public Emoji Icon { get; set; } = new();
+
+    [JsonPropertyName("properties")]
+    public Properties Properties { get; set; } = new();
+
+    [JsonPropertyName("cover")]
+    public Cover Cover { get; set; } = new();
+}
