+++ Experience2Notion/Models/Notions/NotionPageCreate.cs
@@ -1,0 +1,5 @@
+﻿namespace Experience2Notion.Models.Notions;
+public class NotionPageCreate : NotionPageBase
+{
+
+}
