+++ Experience2Notion/Models/Notions/Properties/External.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions.Properties;
+public class External
+{
+    [JsonPropertyName("url")]
+    public string Url { get; set; } = string.Empty;
+}
