+++ Experience2Notion/Models/Notions/Properties/MultiSelectProperty.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions.Properties;
+public class MultiSelectProperty
+{
+    [JsonPropertyName("multi_select")]
+    public SelectOption[] MultiSelect { get; set; } = [];
+}
