+++ Experience2Notion/Models/Notions/Properties/RichTextProperty.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions.Properties;
+public class RichTextProperty
+{
+    [JsonPropertyName("rich_text")]
+    public TextObject[] RichText { get; set; } = [];
+}
+
