+++ Experience2Notion/Models/Notions/Properties/StatusProperty.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions.Properties;
+public class StatusProperty
+{
+    [JsonPropertyName("status")]
+    public StatusValue Status { get; set; } = new StatusValue();
+}
