+++ Experience2Notion/Models/Notions/Properties/TextObject.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions.Properties;
+public class TextObject
+{
+    [JsonPropertyName("text")]
+    public TextContent Text { get; set; } = new TextContent();
+}
