+++ Experience2Notion/Models/Notions/Properties/TitleProperty.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Experience2Notion.Models.Notions.Properties;
+public class TitleProperty
+{
+    [JsonPropertyName("title")]
+    public TextObject[] Title { get; set; } = [];
+}
